Removed inl files as makeshift libs and put them into utils.
Numerous fixes and enhancements for
BennyHill

+++ rococo.sexy.ide/sexy.ide.debugger.cpp
@@ -151,7 +151,7 @@ namespace
 
                   if (debugControl)
                   {
-                     debugControl->RefreshAtDepth(requiredDepth - 1);
+                     debugControl->RefreshAtDepth((int32)(requiredDepth - 1));
                   }
                }
             }
